Revert "Hw 3"
