+++ data/users.json
@@ -1,1 +1,1 @@
-[{"userId":"5d2dd70eca02155aacf1ee01","name":"Edna Dennis","mail":"b5248267@urhen.com"},{"userId":"5d2dd70eb74a864f54425b3a","name":"Marsha Ratliff","mail":"sit@gmail.com"},{"userId":"5d2dd70eea1afbc8824d3120","name":"Obrien Mays","mail":"in@gmail.com"},{"userId":"5d2dd70e24d36d63e2581f05","name":"Bernice Peterson","mail":"ex@gmail.com"},{"userId":"5d2dd70e175cfc0b9a751f39","name":"Castro Howard","mail":"consequat@gmail.com"},{"userId":"5d2dd70ecb74de0e4bad34da","name":"Rochelle Decker","mail":"deserunt@gmail.com"},{"userId":"5d2dd70e8a3709ef7f386131","name":"Mcguire Byrd","mail":"cillum@gmail.com"},{"userId":"5d2dd70e01fb01c94a6667d4","name":"Deanne Rojas","mail":"cupidatat@gmail.com"}]
+[{"userId":"5d2dd70eca02155aacf1ee01","name":"Edna Dennis","mail":"consequat@gmail.com"},{"userId":"5d2dd70eb74a864f54425b3a","name":"Marsha Ratliff","mail":"sit@gmail.com"},{"userId":"5d2dd70eea1afbc8824d3120","name":"Obrien Mays","mail":"in@gmail.com"},{"userId":"5d2dd70e24d36d63e2581f05","name":"Bernice Peterson","mail":"ex@gmail.com"},{"userId":"5d2dd70e175cfc0b9a751f39","name":"Castro Howard","mail":"consequat@gmail.com"},{"userId":"5d2dd70ecb74de0e4bad34da","name":"Rochelle Decker","mail":"deserunt@gmail.com"},{"userId":"5d2dd70e8a3709ef7f386131","name":"Mcguire Byrd","mail":"cillum@gmail.com"},{"userId":"5d2dd70e01fb01c94a6667d4","name":"Deanne Rojas","mail":"cupidatat@gmail.com"}]
