+++ package.json
@@ -4,7 +4,7 @@
   "description": "node js",
   "main": "index.js",
   "scripts": {
-    "start": "npx nodemon --experimental-worker ./server.js",
+    "start": "npx nodemon ./server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -20,15 +20,11 @@
   "dependencies": {
     "async-json-parse": "^1.0.3",
     "body-parser": "^1.19.0",
-    "cron": "^1.7.1",
     "express": "^4.17.1",
     "fs": "0.0.1-security",
-    "md5": "^2.2.1",
     "moment": "^2.24.0",
-    "node-fetch": "^2.6.0",
     "swagger-ui-express": "^4.0.7",
     "uuid": "^3.3.2",
-    "worker-threads": "^1.0.0",
     "yamljs": "^0.3.0"
   },
   "devDependencies": {
