--- src/utils/cron.js
@@ -1,10 +1,0 @@
-const { CronJob } = require('cron');
-
-module.exports = (worker) => {
-    try {
-        const cronJob = new CronJob('* * * * * *', worker);
-        cronJob.start();
-    } catch (error) {
-        console.log(`${error}`);
-    }
-}
