--- src/utils/worker.js
@@ -1,5 +1,0 @@
-
-const statistics = require('./statistics.js');
-const { parentPort, workerData } = require('worker_threads');
-
-parentPort.postMessage(statistics(workerData));
