+++ swagger.yaml
@@ -122,27 +122,6 @@ paths:
           description: Success
         400:
           description: Invalid status value
-  /posts/statistics:
-    get:
-      tags:
-      - posts
-      summary: Get statistics
-      consumes:
-      - application/json
-      responses:
-        200:
-          description: Success
-          schema:
-            type: object
-            properties:
-              postsPerDay:
-                type: number
-              postsPerWeek:
-                type: number
-              postsPerMonth:
-                type: number
-        400:
-          description: 'Invalid status value'
   /user:
     get:
       tags:
@@ -163,7 +142,7 @@ paths:
     get:
       tags:
       -  user
-      summary: Get user by id and gravatar info
+      summary: Get user by id
       produces:
       -  application/json
       parameters:
@@ -175,16 +154,9 @@ paths:
         200:
           description: Success
           schema:
-            type: object
-            properties:
-              userId:
-                type: string
-              name:
-                type: string
-              mail:
-                type: string
-              gravatarInfo:
-                type: object
+            type: array
+            items:
+              $ref: '#/definitions/user'
         400:
           description: 'Invalid status value'
     delete:
